feat(linter): markdownlint 公式ルールカタログを追加 (v0.19.0 Task 1)
- markdown.rs を markdown/mod.rs と markdown/types.rs に分割（コーディングルール準拠）
- RuleParityStatus (Official/Experimental/Hidden) を新設
- OfficialRuleMeta struct を追加（公式コード・タイトル・説明・docs URL・parity status）
- MarkdownDiagnostic に official_meta フィールドを追加
- MarkdownRule trait に official_meta() メソッドを追加
- HeadingStructureRule を HeadingIncrementRule (MD001) にリネームし公式コード化
- BrokenLinkRule を Hidden (official_meta = None) に設定
- HeadingStructureRule として型エイリアスで後方互換性を維持
- parity regression テストを追加（MD001 violation/valid fixture corpus）
